Performance Optiomization
- Concurrency at Plugin call and  containers to callback changes.
- Cell Generator is fixed.
- HSearchCompiler is not fixed.

+++ src/java/com/bizosys/hsearch/byteutils/SortedBytesChar.java
@@ -52,8 +52,8 @@ public final class SortedBytesChar extends SortedBytesBase<Byte>{
 	@Override
 	protected int compare(byte[] inputB, int offset, Byte matchNo) {
 		byte val = inputB[offset];
-		if ( val == matchNo) return 0;
-		if (val > matchNo) return 1;
+		if ( val == matchNo.byteValue()) return 0;
+		if (val > matchNo.byteValue()) return 1;
 		return -1;
 		
 	}
